refactor: Extract scroll logic into useScrollDetection hook and improve logic
